backend/requirements.txt
# Core
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Auth
djangorestframework-simplejwt==5.3.0
social-auth-app-django==5.4.0
dj-rest-auth==5.0.1
django-allauth==0.58.2

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# WebSockets / Async
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# Security
argon2-cffi==23.1.0
django-ratelimit==4.1.0

# Payments
requests==2.31.0

# Email
django-sendgrid-v5==1.2.3

# API docs
drf-spectacular==0.26.5

# Utilities
Pillow==10.1.0
django-filter==23.5
django-storages==1.14.2
python-decouple==3.8
shortuuid==1.0.11

# Testing
factory-boy==3.3.0
faker==20.1.0

# Monitoring
sentry-sdk==1.38.0
